Fixed some code parts that produced compiler warnings
* Initialize potentionally uninitialized variables to zero
* Made implicit type casts explicit
* Comment out unused variables
* Changed some variable types

+++ src/core/n64video/rdp/blender.c
@@ -327,7 +327,7 @@ static void blender_init_lut(void)
             if (temp & 0x10)
                 res |= (1 << (7 - k));
         }
-        bldiv_hwaccurate_table[i] = res;
+        bldiv_hwaccurate_table[i] = (uint8_t)res;
     }
 }
