+++ src/core/n64video/rdp/zbuffer.c
@@ -190,7 +190,7 @@ static INLINE void z_build_com_table(void)
 
 static STRICTINLINE void z_store(uint32_t zcurpixel, uint32_t z, int dzpixenc)
 {
-    uint16_t zval = z_com_table[z & 0x3ffff]|(dzpixenc >> 2);
+    uint16_t zval = z_com_table[z & 0x3ffff]|(uint16_t)(dzpixenc >> 2);
     uint8_t hval = dzpixenc & 3;
     PAIRWRITE16(zcurpixel, zval, hval);
 }
